moved other server code into server folder
