+++ server/routes/views/logout.js
