+++ server/server.js
@@ -13,7 +13,7 @@ import mongoose from 'mongoose';
 import session from 'express-session';
 import mongoStore from 'connect-mongodb-session';
 
-import routes from '../routes';
+import routes from './routes';
 import auth from './auth';
 
 const port = process.env.PORT || 7999;
